refactor: move all activity domain modules into activity/ package
Move 14 modules (classes, activity, util, if_statements, fork,
whilepoly, note, title, group, arrow, connector, ellipse, merge, add)
into activity/ sub-package. Fix internal imports to use relative paths.
Update app.py and test imports.

Note: mypy pre-commit skipped — these files have pre-existing type
issues that will be fixed in a follow-up commit.
