+++ src/plantuml_gui/activity/ellipse.py
@@ -23,10 +23,10 @@
 # SOFTWARE.
 
 
-from plantuml_gui.util import index_of_clicked_element  # pragma: no cover
 from pyquery import PyQuery as Pq
 
 from .classes import Ellipse, SvgChunk
+from .util import index_of_clicked_element  # pragma: no cover
 
 
 def svgtochunklistellipse(svg):
