+++ src/plantuml_gui/activity/note.py
